Visualizations Added
Two different plots were created to show the data. First a bar plot that shows that the players subscribed to a game related newsletter have a much higher mean played hours compared to the players who are not subscribed to a newsletter. Then a histogram was created to view the full distribution of hours played across the whole dataset.